Updating the solution only on fixed prescribed time steps
The lss updating scheme is not contractive. The converged solution
depends on the initially chosen dt (timesteps). In order to make the
solution less arbitrary with respect to the discrete timestep
representation, the code is restructured such that it updates the
solution only on fixed prescribed time steps.

The lss iteration now takes a fixed time step distribution tfix as
input. Then, the KKT-Schur solution w and the update for u and dt is
computed as usual. After that, the new solution (u,dt) is interpolated
at the prescribed timesteps.