Piggyback iteration for interpolation-LSS-solver
The piggyback iteration is set up for the modified interpolating
lss-solver. The Adjoint variable corresponds only to the primal u that
is interpolated at a fixed time discretization.

All derivatives have been tested against finite difference and match,
although epsilon had to be chosen relatively big (1e-3) for some
derivatives.

The piggyback iteration is tested with mu = 1.0 and 15 iterations. The
relative residual drop of the adjoint is only 1e-4 (absolute residuum
 is 1e-7 at convergence). The reduced gradient is validated against
 finite differences with eps=1e-6. They match with relative error around
 0.2%.