Projection method for adjoint convergence
Interpolation to a fixed time grid is removed. Instead, the convergence
of the adjoint iteration is now acchieved by projecting the iterates onto
the subspace orthogonal to the eigenspace with eigenvalue 1 of the iteration matrix.

The convergence behavior is very similar to what was acchieved with the interapolation
approach (relative residuum drop 1e-4, absolut residuum at convergence 1e-7, relative
error of reduced gradient 0.2%). But the numerical effort is much less compared to the
interpolation approach.